tests/submission/W3C/domhtml/HTMLObjectElement07.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - OBJECT</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/dimension-attributes.html#dom-dim-height" rel="help">
<link rel="author" title="Netscape">
</head>
<body>
<object data="/support/html/sample.html" height="80" type="text/html" width="550"></object>

<script>
var domtest = async_test("The value of attribute height of the object element is read and checked against the expected value.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vheight;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("object");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vheight = testNode.height;

      assert_equals(vheight,"80");
       
    domtest.done();
}));

</script>
<p>The value of attribute height of the object element is read and checked against the expected value.</p>
<div id="log"></div>
</body></html>
